refactor: SearchModal -> BottomSheetModal 재사용 컴포넌트
